Couleur

+++ ft_client.c
@@ -6,7 +6,7 @@
 /*   By: cjimenez <cjimenez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/07 12:33:37 by cjimenez          #+#    #+#             */
-/*   Updated: 2022/03/07 14:11:34 by cjimenez         ###   ########.fr       */
+/*   Updated: 2022/03/09 15:40:15 by cjimenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ static int  ft_check(char *str)
 
 static int  ft_error(void)
 {
-    ft_printf("Unable to send signal\n");
+    ft_printf("\033[0;31m[ERROR]\033[0m: Unable to reach the server\n");
     exit(0);
 }
 
@@ -50,32 +50,37 @@ static void ft_send(int pid, char *msg)
                 if (kill(pid, SIGUSR1))
                     ft_error();
             }
-            else
+            else if (!(msg[i] >> j & 1))
             {
                 if (kill(pid, SIGUSR2))
                     ft_error();
             }
-            usleep(500);
+            usleep(1000);
             j--;
         }
         i++;
     }
 }
 
-static void ft_success(int sig)
+static void ft_success(int sig, siginfo_t *info, void *context)
 {
+    (void)context;
     if (sig == SIGUSR1)
         return;
     else if (sig == SIGUSR2)
-        ft_printf("\033[32m>>>>>Signal received!<<<<<\033[0m\n");
+        ft_printf("\033[32m[SUCESS]\033[0m: signal sent to \033[0;33m[%d]\033[0m\n", info->si_pid);
 }
 
 int main(int ac, char **av)
 {
+    struct sigaction	sa_usr;
+    
+	sa_usr.sa_flags = SA_SIGINFO;
+	sa_usr.sa_sigaction = ft_success;
     if (ac == 3 && ft_check(av[1]) == 0)
     {
-        signal(SIGUSR1, ft_success);
-        signal(SIGUSR2, ft_success);
+        sigaction(SIGUSR1, &sa_usr, NULL);
+        sigaction(SIGUSR2, &sa_usr, NULL);
         ft_send(ft_atoi(av[1]), av[2]);
 
     }
